conductor.py
import os
import yaml
import numpy as np
import librosa

# --- CONFIG ---
MEGA_DIR = "/mnt/forge/audio/mega"
FPS = 24

class Conductor:
    def __init__(self):
        self.assets = []
        self.vectors = []

    def scan_census(self):
        print(f"📡 CONDUCTOR: Indexing {MEGA_DIR}...")
        # Recursive scan to catch all 2,273 assets
        for root, _, files in os.walk(MEGA_DIR):
            for f in sorted(files):
                if f.endswith('.wav'):
                    self.assets.append(os.path.join(root, f))

    def generate_library(self):
        """Placeholder for actual latent generation - currently mocks the 128-D stride"""
        print(f"🧬 CONDUCTOR: Synthesizing Latent Library for {len(self.assets)} assets...")
        legend = []
        
        for i, path in enumerate(self.assets):
            # 1. Feature Extraction (Simulated for this pass)
            # In production, this is where we'd run the actual analysis
            v = np.random.rand(128) 
            self.vectors.append(v)
            
            # 2. Shorthand ID Generation (8E0+0p style)
            shorthand = f"{hex(int(v[0]*1000))[2:].upper()}+{int(v[1]*100)}p"
            
            legend.append({
                'id': shorthand,
                'path': path
            })

        # Atomic Write
        np.save("latent_library.npy", np.array(self.vectors))
        with open("legend.yml", "w") as f:
            yaml.dump(legend, f, default_flow_style=False)
        print("✅ Census Complete: .npy and .yml are perfectly synced.")

    def analyze_soundtrack(self, st_path):
        """The Look-Back logic for Suckback detection"""
        print(f"🌊 CONDUCTOR: Analyzing {st_path} for temporal opportunities...")
        # This will populate spine.yml with complex fingerprints (Integrity +1)
        pass

if __name__ == "__main__":
    c = Conductor()
    c.scan_census()
    c.generate_library()