requirements.txt
numpy==1.26.4
librosa==0.10.1
pyyaml==6.0.1
scipy==1.12.0
soundfile==0.12.1